# Conflicts:
#	flying-saucer-core/pom.xml
#	flying-saucer-core/src/main/java/org/xhtmlrenderer/css/constants/IdentValue.java
#	flying-saucer-core/src/main/java/org/xhtmlrenderer/css/constants/Idents.java
#	flying-saucer-core/src/main/java/org/xhtmlrenderer/css/constants/ValueConstants.java
#	flying-saucer-core/src/main/java/org/xhtmlrenderer/css/newmatch/CascadedStyle.java
#	flying-saucer-core/src/main/java/org/xhtmlrenderer/css/newmatch/Selector.java
#	flying-saucer-core/src/main/java/org/xhtmlrenderer/css/parser/CSSParser.java
#	flying-saucer-core/src/main/java/org/xhtmlrenderer/css/parser/FSFunction.java
#	flying-saucer-core/src/main/java/org/xhtmlrenderer/css/parser/FSRGBColor.java
#	flying-saucer-core/src/main/java/org/xhtmlrenderer/css/parser/property/Conversions.java
#	flying-saucer-core/src/main/java/org/xhtmlrenderer/css/parser/property/PrimitivePropertyBuilders.java
#	flying-saucer-core/src/main/java/org/xhtmlrenderer/css/sheet/FontFaceRule.java
#	flying-saucer-core/src/main/java/org/xhtmlrenderer/css/style/CalculatedStyle.java
#	flying-saucer-core/src/main/java/org/xhtmlrenderer/extend/OutputDevice.java
#	flying-saucer-core/src/main/java/org/xhtmlrenderer/layout/BoxBuilder.java
#	flying-saucer-core/src/main/java/org/xhtmlrenderer/layout/Layer.java
#	flying-saucer-core/src/main/java/org/xhtmlrenderer/layout/SharedContext.java
#	flying-saucer-core/src/main/java/org/xhtmlrenderer/render/AbstractOutputDevice.java
#	flying-saucer-core/src/main/java/org/xhtmlrenderer/render/BorderPainter.java
#	flying-saucer-core/src/main/java/org/xhtmlrenderer/swing/Java2DOutputDevice.java
#	flying-saucer-core/src/main/java/org/xhtmlrenderer/util/GeneralUtil.java
#	flying-saucer-examples/pom.xml
#	flying-saucer-fop/pom.xml
#	flying-saucer-log4j/pom.xml
#	flying-saucer-pdf-itext5/pom.xml
#	flying-saucer-pdf-itext5/src/main/java/org/xhtmlrenderer/pdf/ITextOutputDevice.java
#	flying-saucer-pdf-osgi/pom.xml
#	flying-saucer-pdf/pom.xml
#	flying-saucer-pdf/src/main/java/org/xhtmlrenderer/pdf/ITextFontResolver.java
#	flying-saucer-pdf/src/main/java/org/xhtmlrenderer/pdf/ITextOutputDevice.java
#	flying-saucer-swt-examples/pom.xml
#	flying-saucer-swt/pom.xml
#	flying-saucer-swt/src/main/java/org/xhtmlrenderer/swt/SWTOutputDevice.java
#	pom.xml

+++ flying-saucer-core/src/main/java/org/xhtmlrenderer/util/IOUtil.java
@@ -1,56 +1,56 @@
 package org.xhtmlrenderer.util;
 
-import java.io.*;
+import com.google.errorprone.annotations.CheckReturnValue;
+import org.jspecify.annotations.Nullable;
+
+import java.io.BufferedInputStream;
+import java.io.BufferedOutputStream;
+import java.io.ByteArrayOutputStream;
+import java.io.Closeable;
+import java.io.File;
+import java.io.FileNotFoundException;
+import java.io.IOException;
+import java.io.InputStream;
+import java.io.OutputStream;
+import java.net.MalformedURLException;
 import java.net.URL;
 import java.net.URLConnection;
+import java.nio.file.Path;
+import java.util.logging.Level;
+
+import static java.nio.file.Files.newInputStream;
+import static java.nio.file.Files.newOutputStream;
 
 /**
  * @author patrick
  */
 public class IOUtil {
-    public static File copyFile(File page, File outputDir) throws IOException {
-        InputStream in = null;
-        OutputStream out = null;
-        try {
-            in = new BufferedInputStream(new FileInputStream(page));
-            File outputFile;
-            outputFile = new File(outputDir, page.getName());
-            out = new BufferedOutputStream(new FileOutputStream(outputFile));
+    public static void copyFile(File page, File outputDir) throws IOException {
+        File outputFile = new File(outputDir, page.getName());
 
-            // Transfer bytes from in to out
-            byte[] buf = new byte[1024];
-            int len;
-            while ((len = in.read(buf)) > 0) {
-                out.write(buf, 0, len);
+        try (OutputStream out = new BufferedOutputStream(newOutputStream(outputFile.toPath()))) {
+            try (InputStream in = new BufferedInputStream(newInputStream(page.toPath()))) {
+                copyBytes(in, out);
             }
-            in.close();
             out.flush();
-            out.close();
-            return outputFile;
-        } finally {
-            if (in != null) {
-                try {
-                    in.close();
-                } catch (IOException e) {
-                    // swallow
-                }
-            }
-            if (out != null) {
-                try {
-                    out.close();
-                } catch (IOException e) {
-                    // swallow
-                }
-            }
+        }
+    }
+
+    private static void copyBytes(InputStream in, OutputStream out) throws IOException {
+        byte[] buf = new byte[1024];
+        int len;
+        while ((len = in.read(buf)) >= 0) {
+            out.write(buf, 0, len);
         }
     }
 
     public static void deleteAllFiles(final File dir) throws IOException {
         File[] files = dir.listFiles();
-        for (int i = 0; i < files.length; i++) {
-            File file = files[i];
-            if (!file.delete()) {
-                throw new IOException("Cleanup directory " + dir + ", can't delete file " + file);
+        if (files != null) {
+            for (File file : files) {
+                if (!file.delete()) {
+                    throw new IOException("Cleanup directory " + dir + ", can't delete file " + file);
+                }
             }
         }
     }
@@ -60,36 +60,84 @@ public class IOUtil {
      * and reading from it. will return the stream, or null if unable to open or read or a timeout occurred. Does not
      * buffer the stream.
      */
+    @Nullable
     public static InputStream openStreamAtUrl(String uri) {
-        InputStream is = null;
         try {
-            final URLConnection uc = new URL(uri).openConnection();
-
-            // If using Java 5+ you can set timeouts for the URL connection--useful if the remote
-            // server is down etc.; the default timeout is pretty long
-            //
-            //uc.setConnectTimeout(10 * 1000);
-            //uc.setReadTimeout(30 * 1000);
-            //
-            // TODO:CLEAN-JDK1.4
-            // Since we target 1.4, we use a couple of system properties--note these are only supported
-            // in the Sun JDK implementation--see the Net properties guide in the JDK
-            // e.g. file:///usr/java/j2sdk1.4.2_17/docs/guide/net/properties.html
-            System.setProperty("sun.net.client.defaultConnectTimeout", String.valueOf(10 * 1000));
-            System.setProperty("sun.net.client.defaultReadTimeout", String.valueOf(30 * 1000));
-
+            URLConnection uc = new URL(uri).openConnection();
+            uc.setConnectTimeout(10 * 1000);
+            uc.setReadTimeout(30 * 1000);
+            uc.setRequestProperty("Accept", "*/*");
             uc.connect();
 
-            is = uc.getInputStream();
-        } catch (java.net.MalformedURLException e) {
+            return uc.getInputStream();
+        } catch (MalformedURLException e) {
             XRLog.exception("bad URL given: " + uri, e);
         } catch (FileNotFoundException e) {
-            XRLog.exception("item at URI " + uri + " not found");
+            XRLog.exception("item at URI " + uri + " not found (caused by: " + e + ")");
         } catch (IOException e) {
             XRLog.exception("IO problem for " + uri, e);
         }
 
-        return is;
+        return null;
+    }
 
+    /**
+     * Gets a Reader for the resource identified
+     *
+     * @return The stylesheet value
+     */
+    @Nullable
+    @CheckReturnValue
+    public static InputStream getInputStream(@Nullable String uri) {
+        if (uri == null) return null;
+        try {
+            return new BufferedInputStream(new URL(uri).openStream());
+        } catch (MalformedURLException e) {
+            XRLog.exception("bad URL given: " + uri, e);
+        } catch (FileNotFoundException e) {
+            XRLog.exception("item at URI " + uri + " not found (caused by: " + e + ")");
+        } catch (IOException e) {
+            XRLog.exception("IO problem for " + uri, e);
+        }
+        return null;
+    }
+
+    @CheckReturnValue
+    public static byte @Nullable [] readBytes(String uri) {
+        try (InputStream is = getInputStream(uri)) {
+            if (is == null) return null;
+            return readBytes(is);
+        } catch (IOException e) {
+            XRLog.load(Level.WARNING, "Unable to read " + uri, e);
+            return null;
+        }
+    }
+
+    @CheckReturnValue
+    public static byte[] readBytes(Path file) throws IOException {
+        try (InputStream is = newInputStream(file)) {
+            return readBytes(is);
+        }
+    }
+
+    @CheckReturnValue
+    public static byte[] readBytes(InputStream is) throws IOException {
+        ByteArrayOutputStream result = new ByteArrayOutputStream(is.available());
+        copyBytes(is, result);
+        return result.toByteArray();
+    }
+
+    /**
+     * @deprecated Use try-with-resources idiom instead.
+     */
+    @Deprecated
+    @SuppressWarnings("EmptyCatch")
+    public static void close(@Nullable Closeable in) {
+        if (in != null) {
+            try {
+                in.close();
+            } catch (IOException ignore) {
+            }
+        }
     }
 }
